improving modularization
