+++ dojot/module/Config.py
