+++ dojot/module/Logger.py
